+++ cmd/list.go
@@ -4,8 +4,6 @@ import (
 	"errors"
 	"os"
 
-	"github.com/web-of-things-open-source/tm-catalog-cli/internal/model"
-
 	"github.com/spf13/cobra"
 	"github.com/web-of-things-open-source/tm-catalog-cli/internal/app/cli"
 	"github.com/web-of-things-open-source/tm-catalog-cli/internal/remotes"
@@ -18,7 +16,8 @@ var listCmd = &cobra.Command{
 	Short: "List TMs in catalog",
 	Long: `List TMs in catalog by name pattern, filters or search. 
 The pattern can be a full name or just a prefix the names shall start with. 
-Name pattern, filters and search can be combined to narrow down the result.`,
+Name pattern, filters and search can be combined to narrow down the result.
+Use --exact to force full-length matching of the name`,
 	Args: cobra.MaximumNArgs(1),
 	Run:  executeList,
 }
@@ -32,6 +31,7 @@ func init() {
 	listCmd.Flags().StringVar(&filterFlags.FilterMpn, "filter.mpn", "", "filter TMs by one or more comma-separated mpn (manufacturer part number)")
 	listCmd.Flags().StringVar(&filterFlags.FilterExternalID, "filter.externalID", "", "filter TMs by one or more comma-separated external ID")
 	listCmd.Flags().StringVarP(&filterFlags.Search, "search", "s", "", "search TMs by their content matching the search term")
+	listCmd.Flags().BoolP("exact", "e", false, "match the TM name exactly. overrides all other search filter flags")
 }
 
 func executeList(cmd *cobra.Command, args []string) {
@@ -42,17 +42,19 @@ func executeList(cmd *cobra.Command, args []string) {
 	if len(args) > 0 {
 		name = args[0]
 	}
-	search := cli.CreateSearchParamsFromCLI(filterFlags, name)
-	if search != nil {
-		search.Options = model.SearchOptions{NameFilterType: model.PrefixMatch}
-	}
-
 	spec, err := remotes.NewSpec(remoteName, dirName)
 	if errors.Is(err, remotes.ErrInvalidSpec) {
 		cli.Stderrf("Invalid specification of target repository. --remote and --directory are mutually exclusive. Set at most one")
 		os.Exit(1)
 	}
 
+	exact, err := cmd.Flags().GetBool("exact")
+	if err != nil {
+		cli.Stderrf("invalid --exact flag")
+		os.Exit(1)
+	}
+
+	search := cli.CreateSearchParamsFromCLI(filterFlags, name, exact)
 	err = cli.List(spec, search)
 	if err != nil {
 		os.Exit(1)