+++ cmd/pull.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/spf13/cobra"
 	"github.com/web-of-things-open-source/tm-catalog-cli/internal/app/cli"
-	"github.com/web-of-things-open-source/tm-catalog-cli/internal/model"
 	"github.com/web-of-things-open-source/tm-catalog-cli/internal/remotes"
 )
 
@@ -17,7 +16,8 @@ var pullCmd = &cobra.Command{
 	Short: "Pull TMs from a catalog.",
 	Long: `Pulls one or more TMs from a catalog by name pattern, filters or search. 
 The pattern can be a full name or just a prefix the names shall start with. 
-Name pattern, filters and search can be combined to narrow down the result.`,
+Name pattern, filters and search can be combined to narrow down the result.
+Use --exact to force full-length matching of the name`,
 	Args: cobra.MaximumNArgs(1),
 	Run:  executePull,
 }
@@ -32,6 +32,7 @@ func init() {
 	pullCmd.Flags().StringVar(&pFilterFlags.FilterMpn, "filter.mpn", "", "filter TMs by one or more comma-separated mpn (manufacturer part number)")
 	pullCmd.Flags().StringVar(&pFilterFlags.FilterExternalID, "filter.externalID", "", "filter TMs by one or more comma-separated external ID")
 	pullCmd.Flags().StringVarP(&pFilterFlags.Search, "search", "s", "", "search TMs by their content matching the search term")
+	pullCmd.Flags().BoolP("exact", "e", false, "match the TM name exactly. overrides all other search filter flags")
 	_ = pullCmd.MarkFlagRequired("output")
 }
 
@@ -50,11 +51,12 @@ func executePull(cmd *cobra.Command, args []string) {
 	if len(args) > 0 {
 		name = args[0]
 	}
-	search := cli.CreateSearchParamsFromCLI(pFilterFlags, name)
-	if search != nil {
-		search.Options = model.SearchOptions{NameFilterType: model.PrefixMatch}
+	exact, err := cmd.Flags().GetBool("exact")
+	if err != nil {
+		cli.Stderrf("invalid --exact flag")
+		os.Exit(1)
 	}
-
+	search := cli.CreateSearchParamsFromCLI(pFilterFlags, name, exact)
 	err = cli.NewPullExecutor(remotes.DefaultManager()).Pull(spec, search, outputPath)
 
 	if err != nil {